+++ ubi-ruby-fips/push_multiarch.sh
@@ -9,26 +9,35 @@ source ../versions.env
 set +a
 
 REGISTRY="$(normalize_repo_name "$1")"
-ARCHITECTURE=$(../resolve_architecture.sh)
 
 # ${1} - image name
 # ${2} - image tag
 # ${3} - tag suffix
 push_image() {
-  LOCAL_IMAGE="${1}:${2}-${ARCHITECTURE}"
-  IMAGE="cyberark/${1}"
   TAG=$(<../VERSION)
+  IMAGE="cyberark/${1}"
 
   if [[ -n "${REGISTRY:-}" ]]; then
     IMAGE="${REGISTRY}${IMAGE}"
   fi
 
   # Push to either internal or public registry depending on whether registry parameter was passed
-  tag_and_push "${LOCAL_IMAGE}" "${IMAGE}:${UBI_VERSION}-${TAG}${3}-${ARCHITECTURE}"
-  tag_and_push "${LOCAL_IMAGE}" "${IMAGE}:${UBI_VERSION}${3}-${ARCHITECTURE}"
-  tag_and_push "${LOCAL_IMAGE}" "${IMAGE}:${TAG}${3}-${ARCHITECTURE}"
-  tag_and_push "${LOCAL_IMAGE}" "${IMAGE}:${2}-${ARCHITECTURE}"
-
+  create_and_push_manifest \
+    "${IMAGE}:${UBI_VERSION}-${TAG}${3}-amd64" \
+    "${IMAGE}:${UBI_VERSION}-${TAG}${3}-arm64" \
+    "${IMAGE}:${UBI_VERSION}-${TAG}${3}"
+  create_and_push_manifest \
+    "${IMAGE}:${UBI_VERSION}${3}-amd64" \
+    "${IMAGE}:${UBI_VERSION}${3}-arm64" \
+    "${IMAGE}:${UBI_VERSION}${3}"
+  create_and_push_manifest \
+    "${IMAGE}:${TAG}${3}-amd64" \
+    "${IMAGE}:${TAG}${3}-arm64" \
+    "${IMAGE}:${TAG}${3}"
+  create_and_push_manifest \
+    "${IMAGE}:${2}-amd64" \
+    "${IMAGE}:${2}-arm64" \
+    "${IMAGE}:${2}"
 }
 
 push_image "ubi-ruby-builder" "latest" ""